+++ integrations/stripe-graphql-js/src/schema/stripe.ts
@@ -7,6 +7,35 @@ import { stripe } from '../utils'
 
 builder.objectType('Stripe', {
   fields: (t) => ({
+    connectedAccounts: t.field({
+      type: 'StripeConnectedAccounts',
+      resolve: async (_parent, _, context) => {
+        const { isAdmin } = context
+
+        if (!isAdmin) throw new GraphQLYogaError('Not allowed')
+
+        const connectedAccounts = await stripe.accounts.list()
+
+        return connectedAccounts
+      }
+    }),
+    connectedAccount: t.field({
+      type: 'StripeConnectedAccount',
+      args: {
+        id: t.arg.string({
+          required: true
+        })
+      },
+      resolve: async (_parent, { id }, context) => {
+        const { isAdmin } = context
+
+        if (!isAdmin) throw new GraphQLYogaError('Not allowed')
+
+        const connectedAccount = await stripe.accounts.retrieve(id)
+
+        return connectedAccount
+      }
+    }),
     customer: t.field({
       type: 'StripeCustomer',
       args: {