+++ packages/react/src/index.ts
@@ -1,4 +1,5 @@
 export * from './client'
+export * from './components'
 export * from './provider'
 export * from './useAccessToken'
 export * from './useAddSecurityKey'