+++ packages/react/src/useUserRoles.ts
@@ -14,9 +14,10 @@ import { useAuthInterpreter } from './useAuthInterpreter'
  */
 export const useUserRoles = () => {
   const service = useAuthInterpreter()
-  return useSelector(
-    service,
-    (state) => state.context.user?.roles || [],
-    (a, b) => a.every((i) => b.includes(i) && b.every((i) => a.includes(i)))
-  )
+  return useSelector(service, (state) => {
+    if (!state.matches('authentication.signedIn')) {
+      return []
+    }
+    return state.context.user?.roles || []
+  })
 }