+++ dashboard/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @nhost/dashboard
 
+## 0.4.2
+
+### Patch Changes
+
+- 89bd37bc: fix(dashboard): correct redirect URL input opacity
+- Updated dependencies [4601d84e]
+- Updated dependencies [843087cb]
+  - @nhost/react@0.15.0
+  - @nhost/nextjs@1.9.0
+  - @nhost/react-apollo@4.9.0
+
+## 0.4.1
+
+### Patch Changes
+
+- 766cb612: fix(dashboard): correct redirect URL for oauth providers
+- Updated dependencies [53bdc294]
+- Updated dependencies [f2aaff05]
+  - @nhost/nextjs@1.8.3
+  - @nhost/core@0.9.3
+  - @nhost/react@0.14.3
+  - @nhost/nhost-js@1.6.1
+  - @nhost/react-apollo@4.8.3
+
 ## 0.4.0
 
 ### Minor Changes